gmmail表修改

+++ game/src/main/java/cn/game/games/net/data/mapper/GmMailMapper.java
@@ -11,55 +11,68 @@ public interface GmMailMapper {
 	/**
 	 * @mbg.generated
 	 */
-	int deleteByPrimaryKey(int id);
+	int deleteByPrimaryKey(long id);
+
 	/**
 	 * @mbg.generated
 	 */
 	int insert(GmMail row);
+
 	/**
 	 * @mbg.generated
 	 */
 	int insertOrUpdate(GmMail row);
+
 	/**
 	 * @mbg.generated
 	 */
-	GmMail selectByPrimaryKey(int id);
+	GmMail selectByPrimaryKey(long id);
+
 	/**
 	 * @mbg.generated
 	 */
 	int updateByPrimaryKey(GmMail row);
+
 	/**
 	 * @mbg.generated
 	 */
 	List<GmMail> selectAll();
+
 	/**
 	 * @mbg.generated
 	 */
 	List<GmMail> getBatchOffset(@Param("offset") int offset, @Param("limit") int limit);
+
 	/**
 	 * @mbg.generated
 	 */
-	List<GmMail> getBatchCursor(@Param("lastId") int lastId, @Param("limit") int limit);
+	List<GmMail> getBatchCursor(@Param("lastId") long lastId, @Param("limit") int limit);
+
 	/**
 	 * @mbg.generated
 	 */
 	long getTotal();
+
 	/**
 	 * @mbg.generated
 	 */
 	int insertBatch(List<GmMail> records);
+
 	/**
 	 * @mbg.generated
 	 */
 	int deleteBatch(List<GmMail> records);
+
 	/**
 	 * @mbg.generated
 	 */
 	int updateBatch(@Param("recordList") List<GmMail> recordList);
+
 	/**
 	 * @mbg.generated
 	 */
-	Integer getLastIdOfBatch(@Param("lastId") int lastId, @Param("limit") int limit);
+	Long getLastIdOfBatch(@Param("lastId") long lastId, @Param("limit") int limit);
+
 	/**
 	 * 根据主键动态更新指定字段 <p>通过Map参数指定要更新的字段名和对应的值，只更新Map中包含的字段</p>
 	 * @param id  id（主键）
@@ -68,7 +81,7 @@ public interface GmMailMapper {
 	 * @note   1. 字段名必须与数据库列名一致 2. 字段值类型需要与数据库字段类型兼容 3. 主键字段不会被更新 4. 如果params为空或不包含任何有效字段，将不执行更新操作
 	 * @mbg.generated
 	 */
-	int updateColumnsByPrimaryKey(@Param("id") int id, @Param("params") java.util.Map<String, Object> params);
+	int updateColumnsByPrimaryKey(@Param("id") long id, @Param("params") java.util.Map<String, Object> params);
 
 	List<GmMail> selectGlobalMailList();
